requirements-django42.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements-django42.txt requirements/base.txt requirements/django42.txt requirements/forced-upgrade.txt
#
anyio==4.2.0
    # via httpx
argus-server==1.14.1
    # via -r requirements/base.txt
asgiref==3.7.2
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
    #   django-cors-headers
async-timeout==4.0.3
    # via redis
attrs==23.2.0
    # via
    #   automat
    #   jsonschema
    #   referencing
    #   service-identity
    #   twisted
autobahn==22.12.1
    # via daphne
automat==22.10.0
    # via twisted
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
channels[daphne]==4.0.0
    # via
    #   argus-server
    #   channels
    #   channels-redis
channels-redis==4.2.0
    # via argus-server
charset-normalizer==3.3.2
    # via requests
constantly==23.10.4
    # via twisted
cryptography==42.0.2
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
    #   social-auth-core
daphne==4.1.0
    # via channels
defusedxml==0.8.0rc2
    # via
    #   python3-openid
    #   social-auth-core
dj-database-url==2.1.0
    # via argus-server
django==4.2.10
    # via
    #   -r requirements/django42.txt
    #   argus-server
    #   channels
    #   dj-database-url
    #   django-cors-headers
    #   django-filter
    #   django-multiselectfield
    #   django-phonenumber-field
    #   djangorestframework
    #   drf-rw-serializers
    #   drf-spectacular
    #   social-auth-app-django
django-cors-headers==4.3.1
    # via argus-server
django-filter==23.5
    # via argus-server
django-multiselectfield==0.1.12
    # via argus-server
django-phonenumber-field[phonenumberslite]==7.3.0
    # via argus-server
djangorestframework==3.14.0
    # via
    #   argus-server
    #   drf-spectacular
drf-rw-serializers==1.1.1
    # via argus-server
drf-spectacular==0.27.1
    # via argus-server
exceptiongroup==1.2.0
    # via anyio
factory-boy==3.3.0
    # via argus-server
faker==23.1.0
    # via factory-boy
h11==0.14.0
    # via httpcore
httpcore==1.0.2
    # via httpx
httpx==0.26.0
    # via argus-server
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.6
    # via
    #   anyio
    #   httpx
    #   hyperlink
    #   requests
    #   twisted
incremental==22.10.0
    # via twisted
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.21.1
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
msgpack==1.0.7
    # via channels-redis
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
phonenumberslite==8.13.30
    # via django-phonenumber-field
psycopg2==2.9.9
    # via argus-server
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.3.0
    # via service-identity
pycparser==2.21
    # via cffi
pyjwt==2.8.0
    # via social-auth-core
pymsteams==0.2.2
    # via -r requirements/base.txt
pyopenssl==24.0.0
    # via twisted
python-dataporten-auth==2.0.0
    # via argus-server
python-dateutil==2.8.2
    # via faker
python3-openid==3.2.0
    # via social-auth-core
pytz==2024.1
    # via djangorestframework
pyyaml==6.0.1
    # via drf-spectacular
redis==5.0.1
    # via channels-redis
referencing==0.33.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   pymsteams
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
service-identity==24.1.0
    # via twisted
six==1.16.0
    # via
    #   automat
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
social-auth-app-django==5.4.0
    # via argus-server
social-auth-core==4.5.2
    # via
    #   argus-server
    #   python-dataporten-auth
    #   social-auth-app-django
sqlparse==0.4.4
    # via django
twisted[tls]==23.10.0
    # via
    #   daphne
    #   twisted
txaio==23.1.1
    # via autobahn
typing-extensions==4.9.0
    # via
    #   anyio
    #   asgiref
    #   dj-database-url
    #   twisted
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.2.0
    # via requests
wheel==0.42.0
    # via argus-server
whitenoise==6.6.0
    # via argus-server
zope-interface==6.1
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
